fix(shortpassword): demo拆解与规范
