+++ src/packages/shortpassword/demos/taro/demo4.tsx
@@ -1,0 +1,38 @@
+import React, { useState } from 'react'
+import { Cell, ShortPassword, NumberKeyboard } from '@nutui/nutui-react-taro'
+
+const Demo4 = () => {
+  const [visible4, setVisible4] = useState(false)
+  const [visible, setVisible] = useState(false)
+  const [value, setValue] = useState('')
+  const onChange = (v: string) => {
+    setValue((value) => value + v)
+  }
+  const onDelete = () => {
+    setValue((value) => value.slice(0, -1))
+  }
+  return (
+    <>
+      <Cell title="自定义密码长度: 4" onClick={() => setVisible4(true)} />
+      <ShortPassword
+        visible={visible4}
+        value={value}
+        onFocus={() => setVisible(true)}
+        onClose={() => {
+          setVisible4(false)
+          setValue('')
+        }}
+        onChange={(value) => setValue(value)}
+        onComplete={() => setVisible(false)}
+        length={4}
+      />
+      <NumberKeyboard
+        visible={visible}
+        onClose={() => setVisible(false)}
+        onChange={onChange}
+        onDelete={onDelete}
+      />
+    </>
+  )
+}
+export default Demo4
